requirements.txt
pytest
password-validator
valid8
typeguard
rich
python-dotenv
art